Simplify `normalizeDate` and fix timezone discard
For no apparent reason, on this conditional branch, `normalizeDate` was
taking `date[1]`, already a `String` (say `"2015-03-31"`) and calling:

```javascript
getDateString(new Date(date[1]))
```

What happened then is:
```javascript
new Date(date[1])
// in our case => Mon Mar 30 2015 21:00:00 GMT-0300 (BRT)
```

This is because `new Date(str)` when the `str` has no timezone info will
compensate back to the user's timezone. The function assumes the string
is on `GMT+-0`.

The call to `getDateString(new Date(date[1]))` then returns one day
previous to what it should (say `"2015-03-30"`).

This causes date ranges to be wrong on the west hemisphere. On the east
hemisphere, the problem doesn't happen, because it'll compensate
forwards (to say `Tue Mar 31 2015 03:00:00 GMT+0300 (...)`).